+++ asdf/tests/test_generic_io.py
@@ -12,7 +12,6 @@ import asdf
 from asdf import exceptions
 from asdf import util
 from asdf import generic_io
-from asdf.asdf import is_asdf_file
 from asdf.config import config_context
 
 from . import helpers, create_small_tree, create_large_tree
@@ -231,7 +230,7 @@ def test_streams2():
 
     fd = generic_io.InputStream(buff, 'r')
 
-    x = fd._peek(10)
+    x = fd.peek(10)
     x = fd.read()
     assert len(x) == 60
 
@@ -747,21 +746,6 @@ def test_truncated_reader():
     assert tr.read(99) == (init + content[:95])
     assert tr.read(50) == content[95:105]
     assert tr.read() == b''
-
-
-def test_is_asdf(tmpdir):
-    # test fits
-    fits = pytest.importorskip('astropy.io.fits')
-
-    hdul = fits.HDUList()
-    phdu= fits.PrimaryHDU()
-    imhdu= fits.ImageHDU(data=np.arange(24).reshape((4,6)))
-    hdul.append(phdu)
-    hdul.append(imhdu)
-    path = os.path.join(str(tmpdir), 'test.fits')
-    hdul.writeto(path)
-    assert not is_asdf_file(path)
-    assert is_asdf_file(asdf.AsdfFile())
 
 
 def test_blocksize(tree, tmpdir):