+++ CHANGES.rst
@@ -59,6 +59,10 @@
 
 - Fix behavior of ``exception`` argument to ``GenericFile.seek_until``. [#980]
 
+- Fix issues in file type detection to allow non-seekable input and
+  filenames without recognizable extensions.  Remove the ``asdf.asdf.is_asdf_file``
+  function. [#978]
+
 2.7.4 (unreleased)
 ------------------